Add files via upload
'GAD2_data.zip' contains 2 files: 1) GAD2_data1.csv has the data used in running the binomial tests; and 2) GAD2_data2.csv has the data used in making the stacked histogram for the paper (Figure 2).

'GAD2_code.R' contains the code that ran the binomial tests and the stacked histogram. We may need to check this file over again to make sure everything is clear. Also, it might be worth editing this file to just do the percent frequencies for each gesture_type when making the stacked histogram, which would allow us to have all the data in just the single GAD2_data1.csv file (so we wouldn't need to zip together with GAD2_data2.csv)